pyproject.toml
# Root pyproject.toml.
# SonicMJ uses uv for the mjlab migration environment.  Existing subpackages can
# still be installed directly when needed.

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonicmj"
version = "0.1.0"
requires-python = ">=3.10,<3.11"
dependencies = [
    "gear_sonic",
    "mjlab==1.2.0",
    "hydra-core==1.3.2",
    "wandb",
    "trl==0.28.0",
    "transformers>=4.56.2",
    "accelerate>=1.3.0",
    "tensorboard",
    "joblib",
    "easydict",
    "loguru",
    "scipy==1.15.3",
    "pandas",
    "mujoco-warp>=3.8.0,<3.9",
    "vector-quantize-pytorch==1.15.0",
]

[tool.uv.sources]
gear_sonic = { path = "gear_sonic", editable = true }
mujoco = { index = "mujoco" }
mujoco-warp = { index = "pypi" }

[tool.uv]
override-dependencies = ["mujoco>=3.6.0.dev0"]
environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64' and python_version == '3.10'",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "mujoco"
url = "https://py.mujoco.org"
explicit = true

[tool.setuptools.packages.find]
include = ["sonic_mj*"]

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
    | external_dependencies
    | logs
    | source
    | /(
      | external_dependencies/
    )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
skip = [
    "external_dependencies", 
    "external_dependencies/**",
]
skip_glob = [
    "external_dependencies/*",
]

[tool.pyright]
reportPrivateImportUsage = false

[tool.ruff]
line-length = 115
target-version = "py310"
exclude = [
    "gear_sonic/dexmg",
    "external_dependencies",
    "external_dependencies/**",
    "./external_dependencies/**",
    "*.ipynb",
]


[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
case-sensitive = false
combine-as-imports = true
force-sort-within-sections = true
from-first = false
single-line-exclusions = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true
check_untyped_defs = true
exclude = [
    "external_dependencies", 
]

[[tool.mypy.overrides]]
module = "tests.*"
strict_optional = false

[tool.pytest.ini_options]
testpaths = "decoupled_wbc/tests/"
python_classes = [
  "Test*",
  "*Test"
]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
